move alt allele to own file
